### 2021-06-04 bb92a79
Update paper.bib
Added DOIs
