+++ Paper/paper.bib
@@ -9,7 +9,8 @@
   author={Schmidt, Philip and Reiss, Attila and Duerichen, Robert and Marberger, Claus and Van Laerhoven, Kristof},
   booktitle={Proceedings of the 20th ACM international conference on multimodal interaction},
   pages={400--408},
-  year={2018}
+  year={2018},
+  doi={10.1145/3242969.3242985}
 }
 @phdthesis{healey2000wearable,
   title={Wearable and automotive systems for affect recognition from physiology},
@@ -25,7 +26,8 @@
   number={2},
   pages={156--166},
   year={2005},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tits.2005.848368}
 }
 @inproceedings{kutt2018towards,
   title={Towards the development of sensor platform for processing physiological data from wearable sensors},
@@ -33,7 +35,8 @@
   booktitle={International Conference on Artificial Intelligence and Soft Computing},
   pages={168--178},
   year={2018},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-91262-2_16}
 }
 @inproceedings{gjoreski2016continuous,
   title={Continuous stress detection using a wrist device: in laboratory and real life},
@@ -47,7 +50,8 @@
   author={Zhou, Jianlong and Arshad, Syed Z and Luo, Simon and Yu, Kun and Berkovsky, Shlomo and Chen, Fang},
   booktitle={Proceedings of the 2017 CHI Conference Extended Abstracts on Human Factors in Computing Systems},
   pages={2269--2275},
-  year={2017}
+  year={2017},
+  doi={10.1145/3027063.3053140}
 }
 @article{van2019heartpy,
   title={HeartPy: A novel heart rate algorithm for the analysis of noisy signals},
@@ -56,7 +60,8 @@
   volume={66},
   pages={368--378},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.trf.2019.09.015}
 }
 @article{kleiman2019using,
   title={Using wearable physiological monitors with suicidal adolescent inpatients: Feasibility and acceptability study},
@@ -75,7 +80,8 @@
   volume={2020},
   pages={3729},
   year={2020},
-  organization={NIH Public Access}
+  organization={NIH Public Access},
+  doi={10.24251/hicss.2020.456}
 }
 @article{choi2011development,
   title={Development and evaluation of an ambulatory stress monitor based on wearable sensors},
@@ -85,7 +91,8 @@
   number={2},
   pages={279--286},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.37896/jxu15.3/018}
 }
 @article{tarvainen2002advanced,
   title={An advanced detrending method with application to HRV analysis},
@@ -95,7 +102,8 @@
   number={2},
   pages={172--175},
   year={2002},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/10.979357}
 }
 @article{chang2012wireless,
   title={A wireless accelerometer-based body posture stability detection system and its application for meditation practitioners},
@@ -105,7 +113,8 @@
   number={12},
   pages={17620--17632},
   year={2012},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={10.3390/s121217620}
 }
 @article{kildahl2019identification,
   title={Identification of post-traumatic stress disorder in individuals with autism spectrum disorder and intellectual disability: A systematic review},
@@ -115,7 +124,8 @@
   number={1-2},
   pages={1--25},
   year={2019},
-  publisher={Taylor \& Francis}
[…]
