### 2021-07-02 #3328 @professoralkmin
@osorensen I'm interested (and it is always an honor) to review an article for JOSS. However, I can only start this review next week. Is there a problem?
