### 2021-07-06 #3455 @whedon
[REVIEW]: Biosensor Framework: A C# Library for Affective Computing
**Submitting author:** @wsarce (<a href="http://orcid.org/0000-0003-0819-0710">Walker Arce</a>)
**Repository:** <a href="https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework" target ="_blank">https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework</a>
**Version:** 1.0.0
