### 2021-07-31 #3455 @AustinTSchaffer
**Testing:** I don't see any unit tests or integration tests.

**Installation:** The project mentions NuGet and Unity for installation, though provides few details. The Readme should probably at least link to [the project's page on NuGet](https://www.nuget.org/packages/Biosensor-Framework/), which provides additional help with how to get the package installed with a variety of tools. Also, the package itself is a .NET Standard 2.0 project, but the provided examples are .NET Framework 4.7, which makes it hard to tell if this project only works on Windows, or if it can be used on Mac or Linux.

This work appears to satisfy all of the requirements of acceptance, except for automated tests. Unless automated testing is a strict requirement, and barring minor confusions with installation, I recommend acceptance.
