### 2021-07-31 #3455 @osorensen
Thanks for your review @AustinTSchaffer!

Regarding unit tests, the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) say the following:

> Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works

@wsarce, could you please elaborate on how the reviewers can test the software, and if necessary, add additional unit test?

@wsarce, please also answer the issue @AustinTSchaffer raised about Installation.
