+++ README.md
@@ -32,7 +32,9 @@ For our lab to use biosensor data in our research projects, we needed an interfa
         Microsoft.ML.LightGbm (>= 1.5.5)
 
 ### Installation
-The library can be installed using the NuGet package manager and is listed publicly.  For Unity, downloading the package and importing it through the Unity UI is the best approach.  This includes all of the dependency libraries, a test script, a test scene, and an example output file.  This example performs the same tasks as Example2 and needs to be supplied the E4 Streaming Server exe path, an output path, and your API key for your device.
+The library can be installed using the [NuGet package manager](https://www.nuget.org/packages/Biosensor-Framework/) and is listed publicly.  For Unity, downloading the package and importing it through the Unity UI is the best approach.  This includes all of the dependency libraries, a test script, a test scene, and an example output file.  This example performs the same tasks as Example2 and needs to be supplied the E4 Streaming Server exe path, an output path, and your API key for your device.
+
+`Install-Package Biosensor-Framework -Version 1.0.0`
 
 ### Unity Package
 The included Unity package includes the required DLLs, a test script, and an example output file.
