### 2021-08-01 802aece
Added third example to test core library functionality
