+++ Example3_ComputingWithoutAnE4/Example3.cs
@@ -1,0 +1,74 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Threading;
+using MathNet.Numerics.Statistics;
+using Microsoft.ML;
+
+using MMIVR.BiosensorFramework.Extensions;
+using MMIVR.BiosensorFramework.InputPipeline;
+using MMIVR.BiosensorFramework.MachineLearningUtilities;
+
+namespace Example3_ComputingWithoutAnE4
+{
+    class Program
+    {
+        public static MLContext mlContext;
+        public static ITransformer Model;
+        public static DataViewSchema ModelSchema;
+        public static int WindowSize = 5;
+        // Output and input filepaths
+        public static string WesadDirectory = Path.Combine(Environment.CurrentDirectory, "Dataset");
+        public static string SessionOutputPath = Path.Combine(Environment.CurrentDirectory, "Output");
+        public static string ModelDir = Path.Combine(Environment.CurrentDirectory, "Models");
+
+        static void Main(string[] args)
+        {
+            mlContext = new MLContext();
+            Model = mlContext.Model.Load(Path.Combine(ModelDir, "BinModel.zip"), out ModelSchema);
+            List<Tuple<string, List<double[]>, double[]>> ParsedDataset = DataImport.LoadDataset(WesadDirectory);
+            List<ExtractedMultiFeatures> MultiFeatureSet = new List<ExtractedMultiFeatures>();
+
+            foreach (Tuple<string, List<double[]>, double[]> Dataset in ParsedDataset)
+            {
+                int NumberOfSamples = Dataset.Item2[1].Length / 32;
+                for (int i = 0; i < NumberOfSamples - WindowSize; i += WindowSize)
+                {
+                    List<double> SubjectFeatures = new List<double>();
+                    for (int j = 0; j < WindowSize; j++)
+                    {
+                        SubjectFeatures.AddRange(SignalProcessing.ProcessAccSignal(Dataset.Item2[0].GetSubArray(j * 96, (j + 1) * 96)));
+                        SubjectFeatures.AddRange(SignalProcessing.ProcessAccSignal(Dataset.Item2[1].GetSubArray(j * 32, (j + 1) * 32)));
+                        SubjectFeatures.AddRange(SignalProcessing.ProcessAccSignal(Dataset.Item2[2].GetSubArray(j * 32, (j + 1) * 32)));
+                        SubjectFeatures.AddRange(SignalProcessing.ProcessAccSignal(Dataset.Item2[3].GetSubArray(j * 32, (j + 1) * 32)));
+                    }
+                    SubjectFeatures.AddRange(SignalProcessing.ProcessPpgSignal(Dataset.Item2[4].GetSubArray(i * 64, (i + WindowSize) * 64)));
+                    SubjectFeatures.AddRange(SignalProcessing.ProcessEdaSignal(Dataset.Item2[5].GetSubArray(i * 4, (i + WindowSize) * 4)));
+                    SubjectFeatures.AddRange(SignalProcessing.ProcessTmpSignal(Dataset.Item2[6].GetSubArray(i * 4, (i + WindowSize) * 4)));
+
+                    MultiFeatureSet.Add(new ExtractedMultiFeatures()
+                    {
+                        StressFeatures = SubjectFeatures.ToArray().ToFloat(),
+                        Result = (uint)Dataset.Item3.GetSubArray(i * 700, (i + WindowSize) * 700).Mean(),
+                    });
+
+                    if (i + WindowSize > NumberOfSamples)
+                    {
+                        WindowSize = NumberOfSamples - i;
+                    }
+                }
+            }
[…]
