+++ README.md
@@ -42,7 +42,10 @@ The following command can be used to install the package using the NuGet CLI:
 The included Unity package includes the required DLLs, a test script, and an example output file.
 
 ### Testing Installation
-Once the package has been installed, this repo can be cloned to get the test [datasets](https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework/tree/main/Dataset) and the [trained models](https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework/tree/main/Trained%20Models).  Follow Example 1 to test the functionality of the Empatica E4 TCP server communication and window data gathering.  Follow Example 2 to test the functionality of Microsoft.ML for affective computing tasks with a connected body-worn sensor.  Follow Example 3 to demonstrate the functionality of Microsoft.ML for affective computing tasks without a connected body-worn sensor using the WESAD dataset. 
+Once the package has been installed, this repo can be cloned to get the test [datasets](https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework/tree/main/Dataset) and the [trained models](https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework/tree/main/Trained%20Models).  
+Follow Example 1 to test the functionality of the Empatica E4 TCP server communication and window data gathering.  
+Follow Example 2 to test the functionality of Microsoft.ML for affective computing tasks with a connected body-worn sensor.  
+Follow Example 3 to demonstrate the functionality of Microsoft.ML for affective computing tasks without a connected body-worn sensor using the WESAD dataset. 
 
 ### Basic Usage
 This repository has two example projects.  Example1 shows the basic usage to communicate with the server and pull data off of a biosensor.  Example2 expands Example1 by adding in Microsoft.ML inferencing.  Example2 will be used to explain the operation of the library.
