### 2021-08-01 #3455 @wsarce
@osorensen, I've added a third example project that tests the affective computing tasks without the need for an attached body-worn sensor.   It processes the WESAD dataset that is provided in the [Dataset](https://github.com/Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework/tree/main/Dataset) folder of the repo and reports the results in the Console output.

@AustinTSchaffer, The NuGet package link as well as installation CLI command has been added to the readme.  This project was written to be OS agnostic, .NET Standard 2.0 should run on Windows, Linux, and MacOS.  The examples are .NET Framework 4.7 to utilize the Console and were written on Windows 10.  Instructions for installing Unity package has been added to the readme as well.

Let me know if there are any other concerns.  Thank you for the feedback and your time!
