### 2021-08-02 #3455 @osorensen
@wsarce, the paper is well written, but I noted the following issues which need to be fixed before acceptance.

- On page 1, line 21, there should be a full stop (i.e., ".") at the end of the paragraph.
- On page 1, line 40, and page 2, line 1, please make sure the parenthesized references are in alphabetic order.
- On page 2, line 1, please fix the reference "gjoreski2016continuous".
- On page 2, line 51, please make sure the reference "Jennifer Anne Healey, 2000" has the format (Last name, Year).
- On page 2, line 77, the name is missing from the reference.
- On page 4, lines 121-125: the citation does not appear correctly.
- On page 4, lines 130-132: please add DOI 10.2196/13725.
- On page 5, lines 150-153: please make sure WESAD is written in capital letters. You can fix this by putting curly braces around the word "wesad" in your source .bib file, i.e., "{WESAD}".
