+++ Paper/paper.md
@@ -34,13 +34,13 @@ This library provides the following capabilities:
 - Support for simulating a data stream for debugging and testing applications without an Empatica E4 device
 - Support for parsing open source Empatica E4 datasets and training machine learning models on them (i.e., WESAD [@schmidt2018introducing])
 - Support for interfacing new sensors into the same pipeline
-- Contains models trained on WESAD dataset in the Microsoft.ML framework
+- Contains models trained on WESAD dataset in the Microsoft.ML framework.
 
 # Background
 
 Biosensor Framework is a C# library that handles the full process of gathering biometric data from a body-worn sensor, transforming it into handcrafted feature vectors, and delivering an inferencing result in thirty-five lines of code.  Purpose built to handle the Empatica E4 biosensor, the Empatica provided TCP client is wrapped in a static C# class to provide convenient function calls and error handling to simplify code structure.  With the introduction of Microsoft.ML, the entire pipeline is handled in C# managed code and is fully compatible with the Unity game engine, allowing for affective computing to be easily introduced into existing virtual reality therapy tools.  The decoupled structure of the library also allows for new biosensors to be introduced into the pipeline, requiring only a single code file to be added to retrieve data in a list of doubles.  
 
-Generally, affective computing is concerned with the physiological changes of the subject and how that relates to their internal state.  Most implementations of affective computing look at electrodermal activity (EDA), like a lie detector test, to measure the onset of stress before the subject may even be aware of it [@healey2000wearable; @healey2005detecting; @aqajari2020gsr; @kutt2018towards; @gjoreski2016continuous].   There is a growing body of literature that incorporates the use of multiple body sensors, primarily worn on the wrist, such as photoplethysmography (PPG), three axis acceleration (ACC), and skin temperature (TMP) [@kutt2018towards; gjoreski2016continuous; @zhou2017indexing; @van2019heartpy; @kleiman2019using; @kaczor2020objective; @choi2011development].  As these sensors are miniaturized, they can be easily incorporated into wrist-worn packages and can be fused with other sensor modalities, such as respiratory sensors, eye trackers, or body trackers, to estimate a subject’s emotional state more accurately.
+Generally, affective computing is concerned with the physiological changes of the subject and how that relates to their internal state.  Most implementations of affective computing look at electrodermal activity (EDA), like a lie detector test, to measure the onset of stress before the subject may even be aware of it [@healey2000wearable; @healey2005detecting; @aqajari2020gsr; @kutt2018towards; @gjoreski2016continuous].   There is a growing body of literature that incorporates the use of multiple body sensors, primarily worn on the wrist, such as photoplethysmography (PPG), three axis acceleration (ACC), and skin temperature (TMP) [@kutt2018towards; @gjoreski2016continuous; @zhou2017indexing; @van2019heartpy; @kleiman2019using; @kaczor2020objective; @choi2011development].  As these sensors are miniaturized, they can be easily incorporated into wrist-worn packages and can be fused with other sensor modalities, such as respiratory sensors, eye trackers, or body trackers, to estimate a subject’s emotional state more accurately.
 
[…]
