+++ Paper/paper.bib
@@ -5,7 +5,7 @@
   year={2020}
 }
 @inproceedings{schmidt2018introducing,
-  title={Introducing wesad, a multimodal dataset for wearable stress and affect detection},
+  title={Introducing WESAD, a multimodal dataset for wearable stress and affect detection},
   author={Schmidt, Philip and Reiss, Attila and Duerichen, Robert and Marberger, Claus and Van Laerhoven, Kristof},
   booktitle={Proceedings of the 20th ACM international conference on multimodal interaction},
   pages={400--408},
@@ -71,7 +71,8 @@
   number={9},
   pages={e13725},
   year={2019},
-  publisher={JMIR Publications Inc., Toronto, Canada}
+  publisher={JMIR Publications Inc., Toronto, Canada},
+  doi={10.2196/13725}
 }
 @inproceedings{kaczor2020objective,
   title={Objective measurement of physician stress in the emergency department using a wearable sensor},
