+++ Paper/paper.bib
@@ -14,7 +14,7 @@
 }
 @phdthesis{healey2000wearable,
   title={Wearable and automotive systems for affect recognition from physiology},
-  author={Healey, Jennifer Anne},
+  author={Healey, Jennifer A},
   year={2000},
   school={Massachusetts Institute of Technology}
 }
@@ -76,7 +76,7 @@
 }
 @inproceedings{kaczor2020objective,
   title={Objective measurement of physician stress in the emergency department using a wearable sensor},
-  author={Kaczor, Eric E and Carreiro, Stephanie and Stapp, Joshua and Chapman, Brittany and Indic, Premananda},
+  author={Kaczor, Eric and Carreiro, Stephanie and Stapp, Joshua and Chapman, Brittany and Indic, Premananda},
   booktitle={Proceedings of the... Annual Hawaii International Conference on System Sciences. Annual Hawaii International Conference on System Sciences},
   volume={2020},
   pages={3729},
