+++ Paper/paper.md
@@ -42,7 +42,7 @@ Biosensor Framework is a C# library that handles the full process of gathering b
 
 Generally, affective computing is concerned with the physiological changes of the subject and how that relates to their internal state.  Most implementations of affective computing look at electrodermal activity (EDA), like a lie detector test, to measure the onset of stress before the subject may even be aware of it [@healey2000wearable; @healey2005detecting; @aqajari2020gsr; @kutt2018towards; @gjoreski2016continuous].   There is a growing body of literature that incorporates the use of multiple body sensors, primarily worn on the wrist, such as photoplethysmography (PPG), three axis acceleration (ACC), and skin temperature (TMP) [@kutt2018towards; @gjoreski2016continuous; @zhou2017indexing; @van2019heartpy; @kleiman2019using; @kaczor2020objective; @choi2011development].  As these sensors are miniaturized, they can be easily incorporated into wrist-worn packages and can be fused with other sensor modalities, such as respiratory sensors, eye trackers, or body trackers, to estimate a subject’s emotional state more accurately.
 
-Biosensor Framework incorporates the algorithms used in the WESAD dataset for computing feature vectors for all sensors on the Empatica E4: EDA, PPG, ACC, and TMP [-@schmidt2018introducing].  Additionally, the E4 provides estimations on the heart rate and inter-beat interval, as well as a button for manually tagging the data stream, which has been implemented in a function to mark relevant events.  To detrend the EDA signal, a time-varying, finite-impulse-response, high-pass filter is used [@tarvainen2002advanced].  For detecting major responses in the EDA, the method proposed by Healey is used [@healey2000wearable].  To perform the Fast Fourier Transform, the [DSPLib](https://www.codeproject.com/Articles/1107480/DSPLib-FFT-DFT-Fourier-Transform-Library-for-NET-6) project was incorporated to allow real-time computation of the signal spectra.  Finally, a stillness metric was added to compute a unitless measure of movement from the accelerometer signals [@chang2012wireless].
+Biosensor Framework incorporates the algorithms used in the WESAD dataset for computing feature vectors for all sensors on the Empatica E4: EDA, PPG, ACC, and TMP [@schmidt2018introducing].  Additionally, the E4 provides estimations on the heart rate and inter-beat interval, as well as a button for manually tagging the data stream, which has been implemented in a function to mark relevant events.  To detrend the EDA signal, a time-varying, finite-impulse-response, high-pass filter is used [@tarvainen2002advanced].  For detecting major responses in the EDA, the method proposed by Healey is used [@healey2000wearable].  To perform the Fast Fourier Transform, the [DSPLib](https://www.codeproject.com/Articles/1107480/DSPLib-FFT-DFT-Fourier-Transform-Library-for-NET-6) project was incorporated to allow real-time computation of the signal spectra.  Finally, a stillness metric was added to compute a unitless measure of movement from the accelerometer signals [@chang2012wireless].
 
 # Statement of need
 
