+++ Paper/paper.md
@@ -40,7 +40,7 @@ This library provides the following capabilities:
 
 Biosensor Framework is a C# library that handles the full process of gathering biometric data from a body-worn sensor, transforming it into handcrafted feature vectors, and delivering an inferencing result in thirty-five lines of code.  Purpose built to handle the Empatica E4 biosensor, the Empatica provided TCP client is wrapped in a static C# class to provide convenient function calls and error handling to simplify code structure.  With the introduction of Microsoft.ML, the entire pipeline is handled in C# managed code and is fully compatible with the Unity game engine, allowing for affective computing to be easily introduced into existing virtual reality therapy tools.  The decoupled structure of the library also allows for new biosensors to be introduced into the pipeline, requiring only a single code file to be added to retrieve data in a list of doubles.  
 
-Generally, affective computing is concerned with the physiological changes of the subject and how that relates to their internal state.  Most implementations of affective computing look at electrodermal activity (EDA), like a lie detector test, to measure the onset of stress before the subject may even be aware of it [@healey2000wearable; @healey2005detecting; @aqajari2020gsr; @kutt2018towards; @gjoreski2016continuous].   There is a growing body of literature that incorporates the use of multiple body sensors, primarily worn on the wrist, such as photoplethysmography (PPG), three axis acceleration (ACC), and skin temperature (TMP) [@kutt2018towards; @gjoreski2016continuous; @zhou2017indexing; @van2019heartpy; @kleiman2019using; @kaczor2020objective; @choi2011development].  As these sensors are miniaturized, they can be easily incorporated into wrist-worn packages and can be fused with other sensor modalities, such as respiratory sensors, eye trackers, or body trackers, to estimate a subject’s emotional state more accurately.
+Generally, affective computing is concerned with the physiological changes of the subject and how that relates to their internal state.  Most implementations of affective computing look at electrodermal activity (EDA), like a lie detector test, to measure the onset of stress before the subject may even be aware of it [@aqajari2020gsr; @gjoreski2016continuous; @healey2000wearable; @healey2005detecting; @kutt2018towards].   There is a growing body of literature that incorporates the use of multiple body sensors, primarily worn on the wrist, such as photoplethysmography (PPG), three axis acceleration (ACC), and skin temperature (TMP) [@choi2011development; @gjoreski2016continuous; @kaczor2020objective; @kleiman2019using; @kutt2018towards; @van2019heartpy; @zhou2017indexing].  As these sensors are miniaturized, they can be easily incorporated into wrist-worn packages and can be fused with other sensor modalities, such as respiratory sensors, eye trackers, or body trackers, to estimate a subject’s emotional state more accurately.
 
 Biosensor Framework incorporates the algorithms used in the WESAD dataset for computing feature vectors for all sensors on the Empatica E4: EDA, PPG, ACC, and TMP [@schmidt2018introducing].  Additionally, the E4 provides estimations on the heart rate and inter-beat interval, as well as a button for manually tagging the data stream, which has been implemented in a function to mark relevant events.  To detrend the EDA signal, a time-varying, finite-impulse-response, high-pass filter is used [@tarvainen2002advanced].  For detecting major responses in the EDA, the method proposed by Healey is used [@healey2000wearable].  To perform the Fast Fourier Transform, the [DSPLib](https://www.codeproject.com/Articles/1107480/DSPLib-FFT-DFT-Fourier-Transform-Library-for-NET-6) project was incorporated to allow real-time computation of the signal spectra.  Finally, a stillness metric was added to compute a unitless measure of movement from the accelerometer signals [@chang2012wireless].
 
