### 2021-08-03 #3455 @osorensen
Thank you @wsarce! I noticed that also the paper Gjoreski et al. (2016) is missing a doi. Please add https://doi.org/10.1145/2968219.2968306 to this one.
