+++ Paper/paper.bib
@@ -43,7 +43,8 @@
   author={Gjoreski, Martin and Gjoreski, Hristijan and Lu{\v{s}}trek, Mitja and Gams, Matja{\v{z}}},
   booktitle={proceedings of the 2016 ACM international joint conference on pervasive and ubiquitous computing: Adjunct},
   pages={1185--1193},
-  year={2016}
+  year={2016},
+  doi={https://doi.org/10.1145/2968219.2968306}
 }
 @inproceedings{zhou2017indexing,
   title={Indexing cognitive load using blood volume pulse features},
