+++ Paper/paper.bib
@@ -1,9 +1,3 @@
-@article{aqajari2020gsr,
-  title={GSR Analysis for Stress: Development and Validation of an Open Source Tool for Noisy Naturalistic GSR Data},
-  author={Aqajari, Seyed Amir Hossein and Naeini, Emad Kasaeyan and Mehrabadi, Milad Asgari and Labbaf, Sina and Rahmani, Amir M and Dutt, Nikil},
-  journal={arXiv preprint arXiv:2005.01834},
-  year={2020}
-}
 @inproceedings{schmidt2018introducing,
   title={Introducing WESAD, a multimodal dataset for wearable stress and affect detection},
   author={Schmidt, Philip and Reiss, Attila and Duerichen, Robert and Marberger, Claus and Van Laerhoven, Kristof},
