### 2021-08-03 #3455 @osorensen
@wsarce, please see the error message from @whedon above. The DOI 10.1145/2968219.2968306 probably has an incorrect format in your .bib file.
