### 2021-08-04 f86f746
Create CITATION.cff
