### 2021-08-04 8217c56
Update CITATION.cff
