### 2021-08-05 #3455 @wsarce
@osorensen, alright, I got it working.  The DOI from Zenodo is: 10.5281/zenodo.5161984.  All of the metadata looks correct and a citation file was generated for the repo.  Let me know if anything doesn't look right.  Thank you!
