### 2021-08-05 #3455 @osorensen
@wsarce, I noticed that the title of the zenodo archive is "Munroe-Meyer-Institute-VR-Laboratory/Biosensor-Framework: Biosensor-Framework" whereas the title of the paper is "Biosensor Framework: A C# Library for Affective Computing". Could you please edit the title of the archive so it matches the paper?
